[ALPS02043989] ION : modify ion_drv.h
[Detail] struct ion_sys_get_phys_param, "len" data type use unsigned long

[Solution] struct ion_sys_get_phys_param, "len" data type use unsigned long

[Feature] Others

MTK-Commit-Id: dd25efab58a623dd2f1fabd88f29c9af974ab425

CR-Id: ALPS02043989

+++ drivers/staging/android/ion/mtk/ion_drv.h
@@ -87,7 +87,7 @@ typedef struct ion_sys_get_phys_param {
 		struct ion_handle *kernel_handle;
 	};
 	unsigned int phy_addr;
-	unsigned int len;
+	unsigned long len;
 } ion_sys_get_phys_param_t;
 
 #define ION_MM_DBG_NAME_LEN 16
